+++ changelog.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-Last change: 06-JAN-2026 CEST
+Last change: 07-JAN-2026 CEST
 
 0.9.4
 -----
@@ -14,6 +14,7 @@ Important updates:
 *Small improvements:*
 +++++++++++++++++++++
 - Picasso applies constrained layout to all matplotlib figures
+- SPINNA NND changed the way the plot is displayed
 
 *Bug fixes:*
 ++++++++++++